requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==1.10.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.8.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
compressed-rtf==1.0.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.8
Deprecated==1.2.14
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
IMAPClient==2.1.0
kiwisolver==1.4.4
langchain==0.0.150
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
pdfminer.six==20191110
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.0.9
pypdf==3.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.16
tenacity==8.2.2
textract==1.6.5
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
wrapt==1.15.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
